requirements.txt
asgiref==3.9.1
Django==5.2.5
psycopg==3.2.9
psycopg-binary==3.2.9
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
